+++ src/engraving/libmscore/joinMeasure.cpp
@@ -43,7 +43,8 @@ void Score::cmdJoinMeasure(Measure* m1, Measure* m2)
     }
 
     for (size_t staffIdx = 0; staffIdx < nstaves(); ++staffIdx) {
-        if (m1->isMeasureRepeatGroupWithPrevM(staffIdx) || m2->isMeasureRepeatGroupWithNextM(staffIdx)) {
+        if (m1->isMeasureRepeatGroupWithPrevM(static_cast<int>(staffIdx))
+            || m2->isMeasureRepeatGroupWithNextM(static_cast<int>(staffIdx))) {
             MScore::setError(MsError::CANNOT_SPLIT_MEASURE_REPEAT);
             return;
         }
